- Fixed model generation issues (too much indirection going on!)
git-svn-id: https://svn.ritchielab.psu.edu/svn/projects/biofilter@2050 9e62ecf6-770d-0410-a9d7-e62a33e3d092

+++ trunk/src/knowledge/genegenemodelarchive.cpp
@@ -11,7 +11,7 @@
 
 namespace Knowledge {
 	
-uint GeneGeneModelArchive::minImplicationIndex = 1.0;		///< Default to all
+float GeneGeneModelArchive::minImplicationIndex = 1.0;		///< Default to all
 uint GeneGeneModelArchive::maxModelCount = 10000000;		///< Default to 10 million
 
 void GeneGeneModelArchive::LoadFromArchive(const char *filename, bool useBinary) {
